Interactive version
DB.[ch], QV.[ch], and align.[ch] can now be compiled in an interactive
version by setting the defined constant INTERACTIVE in DB.h.  Rather
than printing an error message to stderr and exiting with code 1, the
interactive version of the routines place an error message in a buffer
EPLACE and return an error code to the caller.