DBdump utility, Mask track type, + developer mods
A new DBdump utility has been added that allows one to list any desired
information about any subset of reads in a database.  The output format
is designed to make reading in by a subsequent program as easy as
possible.

Tracks that are masks, i.e. a list of zero or more intervals for each
read, are now specially marked so programs that takes masks via the -m
option can be checked as such.

For developers, in addition to the mask track coding, you should note
that tracks for the untrimmed DB can now be loaded after the DB has
been trimmed.