DB and DAM import and export upgrades
fasta2DB, quiva2DB, and fasta2DAM can now all import entries
from a redirection of the standard input, either with or without a faux
file name.

DB2fasta, DB2quiva, and DAM2fasta, now sends data imported from the
standard input either to a file with the faux name given on import, or
to the standard output if no name was given on import.