Arrow information now supported by DB
New routines arrow2DB and DB2arrow to add and recreate Arrow
information to/from a DB.

DBdump and DBshow now have options to display the new Arrow information.

A number of small changes throughout the rest of the framework to now
account for the fact that a DB can now either be a sequence DB (S-DB),
Quiver DB (Q-DB), or Arrow DB (A-DB).

DBwipe is a new program that removes all Arrow or Quiver information
from a DB, effectively reverting a Q-DB or A-DB back to an S-DB