DBmv added, tracks can now carry meta-data, IO testing (partial)
 A new command DBmv, that renames a DB and all its ancillary files has
been added.

The .anno file of a track can now carry any number of “extra” meta-data
values appended to its end.  The code base of DB.[ch] has definitions
and routines for reading and writing such information.  Catrack has
been extended so that it properly reduces this information across the
block-annotation when producing a single consolidated track for the
entire database.

While not complete, the codes now check that any IO operation whose
failure would cause the result of the command to be incorrect, is
successful and aborts with message otherwise.  Macros to facilitate
this have been added to DB.h.  The commands not yet re-engineered are
the *2DB and fasta2DAM commands.

More commands like DBshow, DBdump, and Catrack now give a brief summary
of their flags when called with no arguments.